listo inciso g

+++ tpe/microservicio-stop/src/main/java/org/example/microserviciostop/DTO/ParadaDTO.java
@@ -9,10 +9,10 @@ public class ParadaDTO {
     private String direccion;
     private String nombre;
     //para localizacion
-    private  int x;
-    private int y;
+    private  float x;
+    private float y;
 
-    public ParadaDTO(Long id, String direccion, String nombre, int x, int y) {
+    public ParadaDTO(Long id, String direccion, String nombre, float x, float y) {
             this.id = id;
             this.direccion = direccion;
             this.nombre = nombre;
